Chore: Move builder back to grafana/grafana
