+++ pkg/apiserver/rest/dualwriter_test.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"testing"
 
-	playlist "github.com/grafana/grafana/pkg/apis/playlist/v0alpha1"
-	"github.com/grafana/grafana/pkg/infra/kvstore"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -45,15 +43,15 @@ func TestSetDualWritingMode(t *testing.T) {
 		ls := legacyStoreMock{m, l}
 		us := storageMock{m, s}
 
-		kvStore := kvstore.WithNamespace(kvstore.NewFakeKVStore(), 0, "storage.dualwriting."+tt.stackID)
+		kvStore := &fakeNamespacedKV{data: make(map[string]string), namespace: "storage.dualwriting." + tt.stackID}
 
 		p := prometheus.NewRegistry()
-		dw, err := SetDualWritingMode(context.Background(), kvStore, ls, us, playlist.GROUPRESOURCE, tt.desiredMode, p)
+		dw, err := SetDualWritingMode(context.Background(), kvStore, ls, us, "playlist.grafana.app/v0alpha1", tt.desiredMode, p)
 		assert.NoError(t, err)
 		assert.Equal(t, tt.expectedMode, dw.Mode())
 
 		// check kv store
-		val, ok, err := kvStore.Get(context.Background(), playlist.GROUPRESOURCE)
+		val, ok, err := kvStore.Get(context.Background(), "playlist.grafana.app/v0alpha1")
 		assert.True(t, ok)
 		assert.NoError(t, err)
 		assert.Equal(t, val, fmt.Sprint(tt.expectedMode))
@@ -81,4 +79,19 @@ func TestCompare(t *testing.T) {
 			assert.Equal(t, tt.expected, Compare(tt.input, exampleObj))
 		})
 	}
+}
+
+type fakeNamespacedKV struct {
+	namespace string
+	data      map[string]string
+}
+
+func (f *fakeNamespacedKV) Get(ctx context.Context, key string) (string, bool, error) {
+	val, ok := f.data[f.namespace+key]
+	return val, ok, nil
+}
+
+func (f *fakeNamespacedKV) Set(ctx context.Context, key, value string) error {
+	f.data[f.namespace+key] = value
+	return nil
 }
