+++ pkg/registry/apis/dashboardsnapshot/exporter.go
@@ -10,8 +10,8 @@ import (
 	"gocloud.dev/blob"
 	"k8s.io/kube-openapi/pkg/spec3"
 
-	"github.com/grafana/grafana/pkg/apiserver/builder"
 	"github.com/grafana/grafana/pkg/infra/db"
+	"github.com/grafana/grafana/pkg/services/apiserver/builder"
 	"github.com/grafana/grafana/pkg/services/dashboardsnapshots"
 )
