+++ pkg/registry/apis/scope/register.go
@@ -16,8 +16,8 @@ import (
 	"k8s.io/kube-openapi/pkg/validation/spec"
 
 	scope "github.com/grafana/grafana/pkg/apis/scope/v0alpha1"
-	"github.com/grafana/grafana/pkg/apiserver/builder"
 	grafanarest "github.com/grafana/grafana/pkg/apiserver/rest"
+	"github.com/grafana/grafana/pkg/services/apiserver/builder"
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	"github.com/prometheus/client_golang/prometheus"
 )
